% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{captcha_download}
\alias{captcha_download}
\title{Download n captcha images from TJRJ website at once}
\usage{
captcha_download(n, dir = "data-raw/captcha")
}
\arguments{
\item{n}{number of captchas to download}

\item{dir}{character string naming a directory path for writing the image}
}
\value{
list of responses
}
\description{
Download n captcha images from TJRJ website at once
}
