% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collect.R
\name{captcha_download_one}
\alias{captcha_download_one}
\title{Download one captcha image from TJRJ website}
\usage{
captcha_download_one(dir)
}
\arguments{
\item{dir}{character string naming a directory path for writing the image}
}
\value{
response
}
\description{
Download just one captcha image from TJRJ website
}
